\name{FusedLasso-package}
\alias{FusedLasso-package}
\alias{FusedLasso}
\docType{package}
\title{
Solves the fused lasso
}
\description{
Solves the fused lasso for general predictor matrices X with weights on the penalties for 
    \code{lambda1} and \code{lambda2} as well as for general linear
    models (binomial implemented).
}
\details{
\tabular{ll}{
Package: \tab FusedLasso\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-19\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}
In order to use the package look at the function \code{\link{fusedlasso}}.
}
\author{
Holger Hoefling

Maintainer: Holger Hoefling <hhoeflin@gmail.com>
}
\keyword{ package }
\seealso{
    \code{\link{fusedlasso}}
}
