\name{testData}
\Rdversion{1.1}
\alias{testData}
\docType{data}
\title{
Data used for testing the package
}
\description{
A set of predictors, responses and true solutions used in order to test if the package is
    working correctly.
}
\usage{data(testData)}
\format{
A list with components corresponding to predictors, responses, lambdas and solutions for various settings.
}
\keyword{datasets}
